+++ CHANGELOG.md
@@ -25,9 +25,11 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 **Enhancements**
 - REST API fields can now be specified directly with the `api_filters` attribute of any `DatabaseTable` class & fields from mix-ins are automatically added
 - add `archive_fields` attribute that sets the "raw data" for the database table & fields from mix-ins are automatically added
+- accept `TOML` input files in addition to `YAML`
 
 **Refactors**
 - the `website.core_components.filters` module has been absorbed into the `DatabaseTable` class/module
+- yaml input for custom workflows now matches the python input format
 
 **Fixes**
 - fix bug in windows dev env where `simmate run-server` fails to find python path